Moving old GEOS code that requires context to one file
