+++ swri_geometry_util/include/swri_geometry_util/intersection.h
@@ -28,6 +28,9 @@
 // *****************************************************************************
 
 #include <vector>
+#define HAVE_INT64_T_64  # Prevents conflict with OpenCV typedef of int64
+#include <geos_c.h>
+#undef HAVE_INT64_T_64
 #include <opencv2/core/core.hpp>
 
 namespace swri_geometry_util
@@ -93,4 +96,6 @@ namespace swri_geometry_util
   double PolygonIntersectionArea(
       const std::vector<cv::Vec2d>& a,
       const std::vector<cv::Vec2d>& b);
+
+  GEOSGeometry* VectorToPolygon(const std::vector<cv::Vec2d>& v);
 }
