+++ swri_geometry_util/src/intersection.cpp
@@ -240,4 +240,28 @@ namespace swri_geometry_util
     finishGEOS();
     return area;
   }
+
+  GEOSGeometry* VectorToPolygon(const std::vector<cv::Vec2d>& v)
+  {
+    // Create GEOS polygon from vector of verticies. Allocate one extra
+    // element so first and last coordinate are the same, closing the polygon
+    GEOSCoordSequence* coords = GEOSCoordSeq_create(v.size() + 1, 2);
+
+    for (size_t i = 0; i < v.size(); i++)
+    {
+      GEOSCoordSeq_setX(coords, i, v.at(i)[0]);
+      GEOSCoordSeq_setY(coords, i, v.at(i)[1]);
+    }
+
+    // Make the first and last coordinate the same to define a closed polygon
+    GEOSCoordSeq_setX(coords, v.size(), v.front()[0]);
+    GEOSCoordSeq_setY(coords, v.size(), v.front()[1]);
+
+
+    GEOSGeometry* ring = GEOSGeom_createLinearRing(coords);
+    GEOSGeometry* polygon = GEOSGeom_createPolygon(ring, 0, 0);
+    GEOSNormalize(polygon);
+
+    return polygon;
+  }
 }
